requirements.txt
flask
torch
transformers
Pillow
SpeechRecognition
gTTS
pydub
pyttsx3
ultralytics